Cleanup
* Move price pipe
* Remove `selectedCategory` input from CatalogComponent
* Add icon name translation service
* Refactor tests
