+++ src/app/app.component.html
@@ -1,8 +1,7 @@
 <mt-header [title]=title [projectLink]=projectLink></mt-header>
 <div class="main-container">
   <div class="data-calc" *ngIf=catalogLoaded>
-    <mt-catalog class="column-element" [selectedCategory]=selectedCategory
-                (categorySelected)="categorySelected($event)"></mt-catalog>
+    <mt-catalog class="column-element" (categorySelected)="categorySelected($event)"></mt-catalog>
     <mt-items class="column-element" [items]="getItemsForCurrentCategory()"
               [orderItems]="getOrderItemsForCurrentCategory()" (itemSelected)="itemSelected($event)"></mt-items>
     <mt-order class="column-element"></mt-order>
