+++ src/app/catalog/catalog.component.ts
@@ -1,10 +1,11 @@
-import {ChangeDetectionStrategy, Component, EventEmitter, Input, OnDestroy, Output} from "@angular/core";
+import {ChangeDetectionStrategy, Component, EventEmitter, Inject, OnDestroy, Output} from "@angular/core";
 import {Catalog} from "../model/catalog";
 import {Category} from "../model/category";
 
 import {CatalogService} from "../service/catalog-service";
 import {Subject} from "rxjs";
 import {takeUntil} from "rxjs/operators";
+import {IconTranslateService, IconTranslateServiceToken} from "../service/icon-translate.service";
 
 @Component({
   selector: "mt-catalog",
@@ -14,18 +15,26 @@ import {takeUntil} from "rxjs/operators";
 })
 export class CatalogComponent implements OnDestroy {
 
-  @Input() selectedCategory: string;
-
   @Output() categorySelected = new EventEmitter<string>();
+
+  selectedCategory: string;
 
   private catalog: Catalog;
 
   private onDestroy = new Subject<void>();
 
-  constructor(private catalogService: CatalogService) {
+  get categories(): Category[] {
+    return this.catalog.getCategories();
+  }
+
+  constructor(private catalogService: CatalogService,
+              @Inject(IconTranslateServiceToken) private iconTranslateService: IconTranslateService) {
     this.catalogService.getCatalogObservable().pipe(
       takeUntil(this.onDestroy)
-    ).subscribe(catalog => this.catalog = catalog);
+    ).subscribe(catalog => {
+      this.catalog = catalog;
+      this.selectedCategory = this.catalog.firstCategory.id;
+    });
   }
 
   ngOnDestroy(): void {
@@ -34,27 +43,11 @@ export class CatalogComponent implements OnDestroy {
   }
 
   onCategorySelected(categoryId) {
+    this.selectedCategory = categoryId;
     this.categorySelected.emit(categoryId);
   }
 
-  getCategories(): Category[] {
-    return this.catalog.getCategories();
-  }
-
   translateIcon(rawIcon: string): string {
-    switch (rawIcon) {
-      case "icon-coin-dollar":
-        return "fa-dollar-sign";
-      case "icon-user":
-        return "fa-user";
-      case "icon-library":
-        return "fa-balance-scale";
-      case "icon-profile":
-        return "fa-id-card";
-      case "icon-flag":
-        return "fa-flag";
-      default:
-        return "";
-    }
+    return this.iconTranslateService.translateIcon(rawIcon);
   }
 }
