+++ src/app/items/items.component.spec.ts
@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { ItemsComponent } from "./items.component";
 import {MatButtonModule, MatCardModule, MatListModule} from "@angular/material";
-import {PricePipe} from "../price.pipe";
+import {PricePipe} from "../pipe/price.pipe";
 import {By} from "@angular/platform-browser";
 import {Item} from "../model/item";
 import {Component, DebugElement} from "@angular/core";
