+++ src/app/service/icon-translate.service.ts
@@ -1,0 +1,7 @@
+import {InjectionToken} from "@angular/core";
+
+export const IconTranslateServiceToken = new InjectionToken<IconTranslateService>("IconTranslate");
+
+export interface IconTranslateService {
+  translateIcon(dataId: string): string;
+}
